pyproject.toml
[project]
dependencies = [
  "django>=5.2,<6.0",
  "psycopg2-binary>=2.9,<3.0",
  "django-allauth>=65.0,<66.0",
  "djangorestframework>=3.15,<4.0",
  "channels>=4.1,<5.0",
  "channels-redis>=4.2,<5.0",
  "celery>=5.4,<6.0",
  "redis>=5.0,<6.0",
  "pillow>=10.0,<11.0",
  "python-decouple>=3.8,<4.0",
  "django-cors-headers>=4.4,<5.0",
  "exchangelib>=5.4,<6.0",
  "uvicorn>=0.30,<1.0",
  "gunicorn>=22.0,<23.0",
  "whitenoise>=6.7,<7.0",
  "django-filter>=24.0,<25.0",
  "django-guardian>=2.4,<3.0",
  "django-phonenumber-field>=8.0,<9.0",
  "phonenumbers>=8.13,<9.0",
  "django-otp>=1.0,<2.0",
  "qrcode>=7.4,<8.0",
  "django-storages>=1.14,<2.0",
  "boto3>=1.35,<2.0",
]
description = "CRM система с Django, PostgreSQL, WebSockets и Celery"
name = "crmpwp"
readme = "README.md"
requires-python = ">=3.13"
version = "0.1.0"

[project.optional-dependencies]
dev = [
  "black>=24.0,<25.0",
  "isort>=5.13,<6.0",
  "flake8>=7.0,<8.0",
  "mypy>=1.11,<2.0",
  "pytest>=8.0,<9.0",
  "pytest-django>=4.8,<5.0",
  "pytest-cov>=5.0,<6.0",
  "factory-boy>=3.3,<4.0",
  "faker>=26.0,<27.0",
]

[tool.setuptools]
packages = ["crm"]

[tool.black]
line-length = 88
target-version = ['py313']

[tool.isort]
multi_line_output = 3
profile = "black"

[tool.mypy]
disallow_incomplete_defs = true
disallow_untyped_defs = true
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "crm.settings"
addopts = "-v --cov=crm --cov-report=html --cov-report=term-missing"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
